--- dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/application/IFizzBuzzCommand.java
@@ -1,7 +1,0 @@
-package org.hiroshima_arc.application;
-
-import org.hiroshima_arc.domain.model.IFizzBuzz;
-
-public interface IFizzBuzzCommand {
-    IFizzBuzz execute(int number);
-}
