+++ dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/application/fizzbuzz/IFizzBuzzCommand.java
@@ -1,0 +1,7 @@
+package org.hiroshima_arc.application.fizzbuzz;
+
+import org.hiroshima_arc.domain.model.fizzbuzz.IFizzBuzz;
+
+public interface IFizzBuzzCommand {
+    IFizzBuzz execute(int number);
+}
