+++ dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/domain/type/fizzbuzz/FizzBuzzType01.java
@@ -1,4 +1,4 @@
-package org.hiroshima_arc.domain.type;
+package org.hiroshima_arc.domain.type.fizzbuzz;
 
 public class FizzBuzzType01 extends IFizzBuzzType {
 }
