+++ dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/domain/type/fizzbuzz/FizzBuzzType03.java
@@ -1,6 +1,6 @@
-package org.hiroshima_arc.domain.type;
+package org.hiroshima_arc.domain.type.fizzbuzz;
 
-import org.hiroshima_arc.domain.model.FizzBuzzValue;
+import org.hiroshima_arc.domain.model.fizzbuzz.FizzBuzzValue;
 
 public class FizzBuzzType03 extends IFizzBuzzType {
     @Override
